refactor project files
