--- lotus/feature.go
@@ -1,7 +1,0 @@
-package lotus
-
-type FeatureMeta struct {
-	FeatureId          string `json:"feature_id,omitempty"`
-	FeatureName        string `json:"feature_name,omitempty"`
-	FeatureDescription string `json:"feature_description,omitempty"`
-}
